+++ src/swf_typed/_history.py
@@ -2,19 +2,17 @@
 
 import abc
 import enum
+import typing as t
 import datetime
 import functools
 import dataclasses
-import typing as t
 
 from . import _common
 
 if t.TYPE_CHECKING:
     import botocore.client
-    from . import _tasks
-    from . import _workflows
-    from . import _activities
-    from . import _executions
+
+    from . import _activities, _executions, _tasks, _workflows
 
 
 class TimeoutType(str, enum.Enum):
@@ -203,10 +201,10 @@ class ActivityTaskCancelledEvent(Event):
     task_started_event_id: int
     """Task start event ID."""
 
-    task_cancel_requested_event_id: int = None
+    task_cancel_requested_event_id: t.Union[int, None] = None
     """Task cancellation request event ID."""
 
-    details: str = None
+    details: t.Union[str, None] = None
     """Cancellation details, usually for explanation."""
 
     @classmethod
@@ -234,7 +232,7 @@ class ActivityTaskCompletedEvent(Event):
     task_started_event_id: int
     """Task start event ID."""
 
-    task_result: str = None
+    task_result: t.Union[str, None] = None
     """Task result."""
 
     @classmethod
@@ -261,10 +259,10 @@ class ActivityTaskFailedEvent(Event):
     task_started_event_id: int
     """Task start event ID."""
 
-    reason: str = None
+    reason: t.Union[str, None] = None
     """Failure reason, usually for classification."""
 
-    details: str = None
+    details: t.Union[str, None] = None
     """Failure details, usually for explanation."""
 
     @classmethod
@@ -298,16 +296,15 @@ class ActivityTaskScheduledEvent(Event):
     decision_event_id: int
     """Task schedule decision event ID."""
 
-    task_input: str = None
+    task_input: t.Union[str, None] = None
     """Task input."""
 
-    control: str = None
+    control: t.Union[str, None] = None
     """Message from task scheduling decider."""
 
     @classmethod
     def from_api(cls, data):
-        from . import _tasks
-        from . import _activities
+        from . import _activities, _tasks
 
         attrs = data["activityTaskScheduledEventAttributes"]
         return cls(
@@ -331,7 +328,7 @@ class ActivityTaskStartedEvent(Event):
     task_scheduled_event_id: int
     """Task schedule event ID."""
 
-    worker_identity: str = None
+    worker_identity: t.Union[str, None] = None
     """Identity of worker which acquired task."""
 
     @classmethod
@@ -360,7 +357,7 @@ class ActivityTaskTimedOutEvent(Event):
     task_started_event_id: int
     """Task start event ID."""
 
-    details: str = None
+    details: t.Union[str, None] = None
     """Most recent progress message from worker (see ``send_heartbeat``)."""
 
     @classmethod
@@ -446,13 +443,12 @@ class ChildWorkflowExecutionCancelledEvent(Event):
     execution_started_event_id: int
     """Child execution start event ID."""
 
-    details: str = None
+    details: t.Union[str, None] = None
     """Cancellation details, usually for explanation."""
 
     @classmethod
     def from_api(cls, data):
-        from . import _workflows
-        from . import _executions
+        from . import _executions, _workflows
 
         attrs = data["childWorkflowExecutionCanceledEventAttributes"]
         return cls(
@@ -484,13 +480,12 @@ class ChildWorkflowExecutionCompletedEvent(Event):
     execution_started_event_id: int
     """Child execution start event ID."""
 
-    execution_result: str = None
+    execution_result: t.Union[str, None] = None
     """Child execution result."""
 
     @classmethod
     def from_api(cls, data):
-        from . import _workflows
-        from . import _executions
+        from . import _executions, _workflows
 
         attrs = data["childWorkflowExecutionCompletedEventAttributes"]
         return cls(
@@ -522,16 +517,15 @@ class ChildWorkflowExecutionFailedEvent(Event):
     execution_started_event_id: int
     """Child execution start event ID."""
 
-    reason: str = None
+    reason: t.Union[str, None] = None
     """Failure reason, usually for classification."""
 
-    details: str = None
+    details: t.Union[str, None] = None
     """Failure details, usually for explanation."""
 
     @classmethod
     def from_api(cls, data):
-        from . import _workflows
-        from . import _executions
+        from . import _executions, _workflows
 
         attrs = data["childWorkflowExecutionFailedEventAttributes"]
         return cls(
@@ -563,8 +557,7 @@ class ChildWorkflowExecutionStartedEvent(Event):
 
     @classmethod
     def from_api(cls, data):
-        from . import _workflows
-        from . import _executions
+        from . import _executions, _workflows
 
         attrs = data["childWorkflowExecutionStartedEventAttributes"]
         return cls(
@@ -596,8 +589,7 @@ class ChildWorkflowExecutionTerminatedEvent(Event):
 
     @classmethod
     def from_api(cls, data):
-        from . import _workflows
-        from . import _executions
+        from . import _executions, _workflows
 
         attrs = data["childWorkflowExecutionTerminatedEventAttributes"]
         return cls(
@@ -630,8 +622,7 @@ class ChildWorkflowExecutionTimedOutEvent(Event):
 
     @classmethod
     def from_api(cls, data):
-        from . import _workflows
-        from . import _executions
+        from . import _executions, _workflows
 
         attrs = data["childWorkflowExecutionTimedOutEventAttributes"]
         assert attrs["timeoutType"] == "START_TO_CLOSE"
@@ -710,7 +701,7 @@ class DecisionTaskCompletedEvent(Event):
     decision_task_started_event_id: int
     """Decision task start event ID."""
 
-    decision_context: str = None
+    decision_context: t.Union[str, None] = None
     """Context provided by decider, accessible when describing execution."""
 
     @classmethod
@@ -737,7 +728,7 @@ class DecisionTaskScheduledEvent(Event):
     decision_task_timeout: t.Union[datetime.timedelta, None] = _common.unset
     """Decision runtime timeout."""
 
-    decision_task_priority: int = None
+    decision_task_priority: t.Union[int, None] = None
     """Decision task priority."""
 
     @classmethod
@@ -764,7 +755,7 @@ class DecisionTaskStartedEvent(Event):
     decision_task_scheduled_event_id: int
     """Decision task schedule event ID."""
 
-    decider_identity: str = None
+    decider_identity: t.Union[str, None] = None
     """Identity of decider which acquired task."""
 
     @classmethod
@@ -894,7 +885,7 @@ class MarkerRecordedEvent(Event):
     decision_event_id: int
     """Marker record decision event ID."""
 
-    details: str = None
+    details: t.Union[str, None] = None
     """Attached marker data."""
 
     @classmethod
@@ -924,13 +915,13 @@ class LambdaFunctionScheduledEvent(Event):
     decision_event_id: int
     """Task schedule event ID."""
 
-    task_input: str = None
+    task_input: t.Union[str, None] = None
     """Lambda function input."""
 
     task_timeout: datetime.timedelta = _common.unset
     """Lambda function invocation timeout."""
 
-    control: str = None
+    control: t.Union[str, None] = None
     """Message from task scheduling decider."""
 
     @classmethod
@@ -981,7 +972,7 @@ class LambdaFunctionCompletedEvent(Event):
     task_started_event_id: int
     """Task start event ID."""
 
-    task_result: str = None
+    task_result: t.Union[str, None] = None
     """Lambda function invocation result."""
 
     @classmethod
@@ -1008,10 +999,10 @@ class LambdaFunctionFailedEvent(Event):
     task_started_event_id: int
     """Task start event ID."""
 
-    reason: str = None
+    reason: t.Union[str, None] = None
     """Failure reason, usually for classification."""
 
-    details: str = None
+    details: t.Union[str, None] = None
     """Failure details, usually for explanation."""
 
     @classmethod
@@ -1134,7 +1125,7 @@ class RequestCancelExternalWorkflowExecutionFailedEvent(Event):
     decision_event_id: int
     """Cancellation request decision event ID."""
 
-    control: str = None
+    control: t.Union[str, None] = None
     """Message from cancellation requesting decider."""
 
     @classmethod
@@ -1165,7 +1156,7 @@ class RequestCancelExternalWorkflowExecutionInitiatedEvent(Event):
     decision_event_id: int
     """Cancellation request decision event ID."""
 
-    control: str = None
+    control: t.Union[str, None] = None
     """Message from cancellation requesting decider."""
 
     @classmethod
@@ -1264,7 +1255,7 @@ class SignalExternalWorkflowExecutionFailedEvent(Event):
     decision_event_id: int
     """External workflow signal decision event ID."""
 
-    control: str = None
+    control: t.Union[str, None] = None
     """Message from signalling decider."""
 
     @classmethod
@@ -1298,10 +1289,10 @@ class SignalExternalWorkflowExecutionInitiatedEvent(Event):
     decision_event_id: int
     """External workflow signal decision event ID."""
 
-    signal_input: str = None
+    signal_input: t.Union[str, None] = None
     """Attached signal data."""
 
-    control: str = None
+    control: t.Union[str, None] = None
     """Message from signalling decider."""
 
     @classmethod
@@ -1324,13 +1315,13 @@ class StartActivityTaskFailedEvent(Event):
 
     type: t.ClassVar[str] = "StartActivityTaskFailed"
 
-    task_id: str = None
+    task_id: t.Union[str, None] = None
     """ID of task to be started."""
 
-    cause: str = None
+    cause: t.Union[str, None] = None
     """Failure cause."""
 
-    task_scheduled_event_id: int = None
+    task_scheduled_event_id: t.Union[int, None] = None
     """Task schedule decision event ID."""
 
     @classmethod
@@ -1352,13 +1343,13 @@ class StartLambdaFunctionFailedEvent(Event):
 
     type: t.ClassVar[str] = "StartLambdaFunctionFailed"
 
-    cause: StartLambdaFailureCause = None
+    cause: t.Union[StartLambdaFailureCause, None] = None
     """Failure cause."""
 
-    message: str = None
+    message: t.Union[str, None] = None
     """Failure explanation."""
 
-    task_scheduled_event_id: int = None
+    task_scheduled_event_id: t.Union[int, None] = None
     """Task schedule decision event ID."""
 
     @classmethod
@@ -1396,13 +1387,12 @@ class StartChildWorkflowExecutionFailedEvent(Event):
     decision_event_id: int
     """Child workflow start decision event ID."""
 
-    control: str = None
+    control: t.Union[str, None] = None
     """Message from child execution starting decider."""
 
     @classmethod
     def from_api(cls, data):
-        from . import _workflows
-        from . import _executions
+        from . import _executions, _workflows
 
         attrs = data["startChildWorkflowExecutionFailedEventAttributes"]
         return cls(
@@ -1439,19 +1429,18 @@ class StartChildWorkflowExecutionInitiatedEvent(Event):
     decision_event_id: int
     """Child execution start decision event ID."""
 
-    execution_input: str = None
+    execution_input: t.Union[str, None] = None
     """Child execution input."""
 
-    execution_tags: t.List[str] = None
+    execution_tags: t.Union[t.List[str], None] = None
     """Child execution tags."""
 
-    control: str = None
+    control: t.Union[str, None] = None
     """Message from child execution starting decider."""
 
     @classmethod
     def from_api(cls, data):
-        from . import _workflows
-        from . import _executions
+        from . import _executions, _workflows
 
         attrs = data["startChildWorkflowExecutionInitiatedEventAttributes"]
         config = _executions.PartialExecutionConfiguration.from_api(attrs)
@@ -1560,7 +1549,7 @@ class TimerStartedEvent(Event):
     decision_event_id: int
     """Timer start decision event ID."""
 
-    control: str = None
+    control: t.Union[str, None] = None
     """Message from timer starting decider."""
 
     @classmethod
@@ -1584,13 +1573,13 @@ class WorkflowExecutionCancelRequestedEvent(Event):
 
     type: t.ClassVar[str] = "WorkflowExecutionCancelRequested"
 
-    cause: ExecutionTerminationCause = None
+    cause: t.Union[ExecutionTerminationCause, None] = None
     """Cancellation request cause."""
 
-    cancelling_execution: "_executions.ExecutionId" = None
+    cancelling_execution: t.Union["_executions.ExecutionId", None] = None
     """Execution which requested the cancellation."""
 
-    cancel_decision_event_id: int = None
+    cancel_decision_event_id: t.Union[int, None] = None
     """Cancellation request decision event ID in execution which requested the
     cancellation.
     """
@@ -1622,7 +1611,7 @@ class WorkflowExecutionCancelledEvent(Event):
     decision_event_id: int
     """Execution cancel decision event ID."""
 
-    details: str = None
+    details: t.Union[str, None] = None
     """Execution cancellation details, usually for explanation."""
 
     @classmethod
@@ -1645,7 +1634,7 @@ class WorkflowExecutionCompletedEvent(Event):
     decision_event_id: int
     """Execution complete decision event ID."""
 
-    execution_result: str = None
+    execution_result: t.Union[str, None] = None
     """Execution result."""
 
     @classmethod
@@ -1679,16 +1668,15 @@ class WorkflowExecutionContinuedAsNewEvent(Event):
     decision_event_id: int
     """Continue as new execution decision event ID."""
 
-    execution_input: str = None
+    execution_input: t.Union[str, None] = None
     """New execution input."""
 
-    execution_tags: t.List[str] = None
+    execution_tags: t.Union[t.List[str], None] = None
     """New execution tags."""
 
     @classmethod
     def from_api(cls, data):
-        from . import _workflows
-        from . import _executions
+        from . import _executions, _workflows
 
         attrs = data["workflowExecutionContinuedAsNewEventAttributes"]
         config = _executions.PartialExecutionConfiguration.from_api(attrs)
@@ -1713,10 +1701,10 @@ class WorkflowExecutionFailedEvent(Event):
     decision_event_id: int
     """Execution fail decision event ID."""
 
-    reason: str = None
+    reason: t.Union[str, None] = None
     """Failure reason, usually for classification."""
 
-    details: str = None
+    details: t.Union[str, None] = None
     """Failure details, usually for explanation."""
 
     @classmethod
@@ -1740,13 +1728,13 @@ class WorkflowExecutionSignaledEvent(Event):
     signal_name: str
     """Signal name."""
 
-    signal_input: str = None
+    signal_input: t.Union[str, None] = None
     """Attached signal data."""
 
-    signalling_execution: "_executions.ExecutionId" = None
+    signalling_execution: t.Union["_executions.ExecutionId", None] = None
     """Execution which sent the signal."""
 
-    signal_decision_event_id: int = None
+    signal_decision_event_id: t.Union[int, None] = None
     """Signal decision event ID in execution which sent the signal."""
 
     @classmethod
@@ -1779,26 +1767,25 @@ class WorkflowExecutionStartedEvent(Event):
     guaranteed).
     """
 
-    execution_input: str = None
+    execution_input: t.Union[str, None] = None
     """Execution input."""
 
-    execution_tags: t.List[str] = None
+    execution_tags: t.Union[t.List[str], None] = None
     """Execution tags."""
 
-    continued_execution_run_id: str = None
+    continued_execution_run_id: t.Union[str, None] = None
     """Run ID of execution which this execution continues from."""
 
-    parent_execution: "_executions.ExecutionId" = None
+    parent_execution: t.Union["_executions.ExecutionId", None] = None
     """Parent execution (which started this execution)."""
 
-    parent_initiated_event_id: int = None
+    parent_initiated_event_id: t.Union[int, None] = None
     """ID of event in parent execution which begins the starting of this
     execution."""
 
     @classmethod
     def from_api(cls, data):
-        from . import _workflows
-        from . import _executions
+        from . import _executions, _workflows
 
         attrs = data["workflowExecutionStartedEventAttributes"]
         config = _executions.PartialExecutionConfiguration.from_api(attrs)
@@ -1827,13 +1814,13 @@ class WorkflowExecutionTerminatedEvent(Event):
     child_execution_policy: "_executions.ChildExecutionTerminationPolicy"
     """Child execution policy (how open child executions were handled)."""
 
-    cause: ExecutionTerminationCause = None
+    cause: t.Union[ExecutionTerminationCause, None] = None
     """Failure cause."""
 
-    reason: str = None
+    reason: t.Union[str, None] = None
     """Termination reason, usually for classification."""
 
-    details: str = None
+    details: t.Union[str, None] = None
     """Termination details, usually for explanation."""
 
     @classmethod
@@ -1898,8 +1885,8 @@ def get_execution_history(
     execution: "_executions.ExecutionId",
     domain: str,
     reverse: bool = False,
-    client: "botocore.client.BaseClient" = None,
-) -> t.Generator[Event, None, None]:
+    client: t.Union["botocore.client.BaseClient", None] = None,
+) -> _common.PageConsumer[Event]:
     """Get workflow execution history; retrieved semi-lazily.
 
     Args:
@@ -1925,7 +1912,7 @@ def get_execution_history(
 def get_last_execution_history_event(
     execution: "_executions.ExecutionId",
     domain: str,
-    client: "botocore.client.BaseClient" = None,
+    client: t.Union["botocore.client.BaseClient", None] = None,
 ) -> Event:
     """Get last workflow execution history event.